+++ plugin-code/src/main/java/com/trivago/files/FileSystemManager.java
@@ -31,6 +31,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -153,7 +154,7 @@ public class FileSystemManager {
      */
     public void writeContentToFile(String content, String filePath) throws FileCreationException {
         try {
-            FileUtils.fileWrite(filePath, "UTF-8", content);
+            FileUtils.fileAppend(filePath, "UTF-8", content);
         } catch (IOException e) {
             throw new FileCreationException(filePath);
         }